+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### v0.21.0.pre
+
+* bug-fixes
+  * Fix  `Products#sort_images` method
+
+* features
+  * Add  `Variations#sort_images` method
 ### v0.20.0.pre
 
 * features